+++ deployment/linux_wheels/wheel_workflow.sh
@@ -16,7 +16,7 @@ source "$HOME/.cargo/env"
 # Rust 1.64.0 will change the minimum glibc ABI
 # to a version incompatible with manylinux_2014,
 # so we need to be careful in general.
-rustup override set 1.62.1 
+rustup override set 1.66.1 
 # Pin cbindgen 
 cargo install --locked cbindgen@0.24.3