+++ fwdpy11/headers/fwdpy11/genetic_value_to_fitness/GaussianStabilizingSelection.hpp
@@ -15,17 +15,18 @@ namespace fwdpy11
 
       public:
         explicit GaussianStabilizingSelection(const GSSmo &input)
-            : GeneticValueIsTrait(input.total_dim), pimpl(input.clone())
+            : GeneticValueIsTrait(), pimpl(input.clone())
         {
         }
 
         explicit GaussianStabilizingSelection(const MultivariateGSSmo &input)
-            : GeneticValueIsTrait(input.total_dim), pimpl(input.clone())
+            : GeneticValueIsTrait(), pimpl(input.clone())
         {
         }
 
         double
-        operator()(const DiploidGeneticValueToFitnessData data) const final {
+        operator()(const DiploidGeneticValueToFitnessData data) const final
+        {
             return this->pimpl->operator()(data);
         }
 
@@ -39,6 +40,12 @@ namespace fwdpy11
         clone() const final
         {
             return this->pimpl->clone();
+        }
+
+        std::size_t
+        ndim() const override
+        {
+            return pimpl->ndim();
         }
     };
 }