+++ fwdpy11/headers/fwdpy11/genetic_value_to_fitness/GeneticValueIsTrait.hpp
@@ -27,8 +27,7 @@ namespace fwdpy11
     struct GeneticValueIsTrait : public GeneticValueToFitnessMap
     /// Another ABC.  Effectively a type trait
     {
-        explicit GeneticValueIsTrait(std::size_t ndim)
-            : GeneticValueToFitnessMap(ndim, maps_to_fitness(false))
+        explicit GeneticValueIsTrait() : GeneticValueToFitnessMap()
         {
         }
         virtual ~GeneticValueIsTrait() = default;
@@ -36,6 +35,10 @@ namespace fwdpy11
         GeneticValueIsTrait(GeneticValueIsTrait&&) = default;
         GeneticValueIsTrait& operator=(const GeneticValueIsTrait&) = delete;
         GeneticValueIsTrait& operator=(GeneticValueIsTrait&&) = default;
+
+        bool is_fitness() const final {
+            return false;
+        }
     };
 } // namespace fwdpy11